requirement.txt
fastapi
strawberry-graphql[fastapi]
uvicorn[standard]
sqlalchemy
sqlalchemy_utils